requirements.txt
asgiref==3.8.1
Django==5.1.6
django-environ==0.12.0
django-simple-history==3.8.0
exceptiongroup==1.2.2
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.4
pytest-django==4.9.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
